Implement basic configuration management.
The configuration is either read from the command line, or, if a config file is given on the command line, then it is read from that file instead.

The config file format is toml.
Since the file-based configuration and the CLI-based configuration is derived from that same struct, the keys are the same.

Relates to #9